Fixes a performance leak and adds granular time logging
The performance leak was caused by fork() running increasing slowly as
the process aged (taking an addition ~30ms / day).

The granular time logging is implemented with a custom profiler class
that tracks the time spent between marked sections in the code.
Annotations have been added to the AoR env and to my stable baselines
fork.