Fix occasional noita launch error.
The error was caused by the mouse being left where noita's main menu
opens, causing noita to select a main menu element prior to our start
game macro running.

The fix is for the env to move the mouse to a safe spot prior to the
game launching. However, we implement this logic in the harness instead
as there's no callback available to the env from the harness with the
correct timing.