Fix env_is_wrapped attribute error crash in Noita training.
SB3 requires an env implement .env_is_wrapped() if it's not
VecFrameMonitored and yet is wrapped with a VecEnvWrapper. We workaround
the issue by using the VecFrameMonitored env for eval. This steps in
tensorboard incomprable with different eval frequencies, which is less
than ideal, but it should get things running.