Check in partially working send_event input backend.
This backend uses XSendEvent to send mouse events and overrides
QueryPointer replies to try to ensure the client sees a consistent
cursor position from the server.

With that said, although the logic appears correct, Noita under
proton seems to crash if overrident QueryPointer replies offset the
cursor position by even a single pixel. This could be due to reporting
incorrect cursor windows, but I haven't gotten the chance to test this.