Use template matching for factorio loading screen env transition.
This is as opposed to just waiting for a fixed
duration for loading to hopfully be finished.
Using template matching instead will be faster
since we don't have to leave any margin of error
and more reliable on slow or bogged down machines
where loading may take longer than expected.